Cleaned up the formatting a bit to make it PEP8 compilant.
Added device registration as per recommended user registration flow.

Added password hash to the user registration.

Explanation on the last point
The user registration flow is as follows:
1. A new user is registered with synapse
2. Their device is registered
3. Element starts negotiating with synapse where to store the user keys.
 3a. The will try and find a main server and a backup.
 3b. The backup fails (since we do not have it) and that is OK
 3c. The main homeserver/synapse will be selected as main key store
 3d. IMPORTANT: Element tries to login wuth username and password
                                            ^^^^^^^^
                to synapse in order to store the keys
 3e. Since synapse didi not have a local password for that username,
     3e FAILS
 3f. An error is displayed by Element saying that the user keys could
     not be stored.

The solution was to add the user hashed password to the user
registration flow. The downside is that every user needs to be
recreated. Also, if somebody
1. Bought valid tickets
2. Logged into matrix.europython.eu
3. Wrote down their Matrix username
4. Reassigned their tickets to somebody else
5. Used a separate client
they could login without having a valid ticket. I think the threat model
is something we can live with. Also, if somebody went through all this
trouble to save ~50 euros, they probably deserve to be able to login.